maybe fix

+++ DiffAugment-stylegan2-pytorch/torch_utils/ops/conv2d_gradfix.py
@@ -142,7 +142,7 @@ def _conv2d_gradfix(transpose, weight_shape, stride, padding, output_padding, di
         def forward(ctx, grad_output, input):
             op = torch._C._jit_get_operation('aten::cudnn_convolution_backward_weight' if not transpose else 'aten::cudnn_convolution_transpose_backward_weight')
             flags = [torch.backends.cudnn.benchmark, torch.backends.cudnn.deterministic, torch.backends.cudnn.allow_tf32]
-            grad_weight = op(weight_shape, grad_output, input, padding, stride, dilation, groups, *flags)
+            grad_weight = op(weight_shape, grad_output.to(input.dtype), input, padding, stride, dilation, groups, *flags)
             assert grad_weight.shape == weight_shape
             ctx.save_for_backward(grad_output, input)
             return grad_weight
